Solve the function as argument problem
- fix an issue where the argument being a ValueFn caused issues with being null.

The ValueFn passed down as an argument is now evaluated as a normal ValueFn and assigned a UUID. When the parameter it applies to is being invoked, the invokation realises that the function with the parameter's name does not exist, and therefore it starts a lookup. In the lookup, it traverses the tree to access the relevant parameters, the uses the parameter's index to access the UUID-named function definition and swaps the failed function definition with the one found in the map.

This solution uses two nasty hacks:
- global, Parser-wide variable with the function aliases map -- it's there to be accessible throughout the run of the program, and
- global, Parser-wide variable with the last Identifier's token vale -- it's a hacky way to get the parameter's name, since the ASTIdentifier method does not return anything by default.