Start implementing 'let' keyword
- add appropriate grammar,
- implement ASTDefinition in Parser,
- start amending all test cases.

TODO: There's a problem with scope in test04 -- when the inner loop finishes executing, the 'j' index variable is not removed from scope, i.e. it is still defined, so on subsequent run it will return an error. Try investigating variable management in scope and try implementing a mechanism for removing variables from the scope level on finished block of code.