Make the cast a bit more robust, make a way for a cast to get the superclass
Whenever there are problems with casting due to a nested class, the new method will attempt to find the superclass using the specified type in the cast. The instance does not change -- Java should take care of handling it, but the type problem caused wrong methods to be found by Reflection, therefore they couldn't be invoked. Now with the right cast to a superclass, it should work correctly.

There's a problem with the current cast however -- InputStreamReader is not cast to Reader -- look into it.